refactor uniswap v3 provider, add default implementation for get_transaction_status
