+++ gemstone/src/swapper/orca/provider.rs
@@ -109,11 +109,6 @@ impl GemSwapProvider for Orca {
     async fn fetch_quote_data(&self, _quote: &SwapQuote, _provider: Arc<dyn AlienProvider>, _data: FetchQuoteData) -> Result<SwapQuoteData, SwapperError> {
         Err(SwapperError::NotImplemented)
     }
-
-    async fn get_transaction_status(&self, _chain: Chain, _transaction_hash: &str, _provider: Arc<dyn AlienProvider>) -> Result<bool, SwapperError> {
-        // TODO: the transaction status from the RPC
-        Ok(true)
-    }
 }
 
 impl Orca {
