+++ gemstone/src/swapper/pancakeswap_aptos/mod.rs
@@ -112,8 +112,4 @@ impl GemSwapProvider for PancakeSwapAptos {
         };
         Ok(data)
     }
-
-    async fn get_transaction_status(&self, _chain: Chain, _transaction_hash: &str, _provider: Arc<dyn AlienProvider>) -> Result<bool, SwapperError> {
-        unimplemented!()
-    }
 }
